+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## neptune-tensorflow-keras 2.0.0
 
 ### Changes
-- `NeptuneCallback` accepts now namespace `Handler` as an alternative to `Run` for the `run` argument. This means that
+- `NeptuneCallback` now accepts namespace `Handler` as an alternative to `Run` for the `run` argument. This means that
   you can call it like `NeptuneCallback(run=run["some/namespace/"])` to log everything to the `some/namespace/` 
   location of the run.